Pulling full genomes using gbwithparts

+++ checkForFeature.py
@@ -4,8 +4,9 @@ import BLAST
 import utilities
 import os
 import time
+from flask import flash
 
-def getFeatureLocation(ids, reference):
+def getFeatureLocation(ids, reference, recordName, recordLocation):
 
     dbpath = "files/temp_blastfiles.fasta"
     querypath = "files/psi.xml"
@@ -26,12 +27,24 @@ def getFeatureLocation(ids, reference):
                 time.sleep(1)
 
             if os.path.isfile(querypath):
-                location = BLAST.getBlastLocation(open(querypath))
-                print (location )
-                # return location
+                blast_info = BLAST.getBlastInfo(open(querypath))
 
-                utilities.removeFile(dbpath, querypath)
-                return location
+                if ("sequence" in blast_info and "location" in blast_info):
+
+                    # Update the record with the new location
+                    setattr(record, recordName, blast_info["sequence"])
+                    setattr(record, recordLocation, blast_info["location"])
+
+                    # Commit the changed record
+                    servers.db.session.add(record)
+                    servers.db.session.commit()
+
+
+                    # Remove created files
+                    utilities.removeFile(dbpath, querypath)
+                else:
+                    flash("Couldn't find an %s region in %s" % (recordName, record.name))
+
         else:
             raise ValueError("%s isn't a file!" % "files/temp_blastfiles.fasta")
 
@@ -46,48 +59,6 @@ def addToDatabase():
     return
 
 
-
-
-# def checkFeature(ids, reference, recordName, recordLocation):
-#
-#     query = models.SequenceRecords.query.filter(models.SequenceRecords.uid.in_(ids))
-#     for record in query.all():
-#         best_align = 0
-#         best_location = None
-#         best_seq = None
-#         print(record.name)
-#         seq_record = servers.bio_db.lookup(primary_id=record.name)
-#
-#         # print(seq_record.description)
-#         # print(seq_record)
-#
-#         for feature in seq_record.features:
-#             # print(feature)
-#             if 'translation' in feature.qualifiers:
-#
-#                 alignment = pairwise2.align.globalms(reference, feature.qualifiers['translation'][0],
-#                                                      2, -1, -1, -.5, score_only=True)
-#                 if alignment > best_align:
-#                     best_align = alignment
-#                     best_seq = feature.qualifiers['translation'][0]
-#                     best_location = feature.location
-#
-#         if (best_seq):
-#             print("best align is ", best_align)
-#             flash("Found an A1 region in %s" % record.name)
-#             location = re.search(r"\d*:\d*", str(best_location))
-#
-#             setattr(record, recordName, best_seq)
-#             print('location is')
-#             print(location)
-#             if location:
-#                 print(location.group(0))
-#                 setattr(record, recordLocation, location.group(0))
-#                 servers.db.session.add(record)
-#                 servers.db.session.commit()
-#
-#         # alignment = pairwise2.align.globalms(yenA1, best_seq,
-#         #                                      2, -1, -1, -.5)
 
 
 def deleteFeature(ids, recordName, recordLocation):
